chore: relocate resolveGateway to storefront checkout utils
