+++ storefronts/features/checkout/utils/resolveGateway.js
@@ -1,17 +1,17 @@
-import resolveGateway from '../../../../core/utils/resolveGateway.js';
+export const SUPPORTED_GATEWAYS = ['stripe', 'authorizeNet', 'paypal', 'nmi', 'segpay'];
 
-export default function getActivePaymentGateway(log = () => {}, warn = () => {}) {
-  const cfg = window.SMOOTHR_CONFIG || {};
+export default function resolveGateway(config = {}, storeSettings = {}) {
+  const provider =
+    config.active_payment_gateway ||
+    storeSettings.active_payment_gateway;
 
-  if (!cfg.active_payment_gateway) {
-    warn('active_payment_gateway not configured');
-    return null;
+  if (!provider) {
+    throw new Error('active_payment_gateway not configured');
   }
 
-  try {
-    return resolveGateway(cfg);
-  } catch (e) {
-    warn('Gateway resolution failed:', e?.message || e);
-    return null;
+  if (!SUPPORTED_GATEWAYS.includes(provider)) {
+    throw new Error(`Unknown payment gateway: ${provider}`);
   }
+
+  return provider;
 }
